+++ blocklib/CMakeLists.txt
@@ -2,6 +2,5 @@ add_subdirectory(helpers)
 add_subdirectory(digitizers)
 
 if (ENABLE_PICOSCOPE)
-    add_subdirectory(fair-picoscope)
-    add_subdirectory(picoscope4000a)
+    add_subdirectory(picoscope)
 endif ()
