+++ blocklib/picoscope/CMakeLists.txt
@@ -5,11 +5,11 @@ set_property(TARGET ps4000a PROPERTY
 target_link_libraries(ps4000a INTERFACE PkgConfig::zlib PkgConfig::libusb)
 target_include_directories(ps4000a INTERFACE ${PICOSCOPE_PREFIX}/include/libps4000a ${PICOSCOPE_PREFIX}/include/libps5000a) # Hack: PicoCallback.h is missing in libps4000a/
 
-add_library(fair-picoscope4000a SHARED picoscope4000a.cc)
-target_include_directories(fair-picoscope4000a INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}> $<INSTALL_INTERFACE:include/>)
+add_library(fair-picoscope INTERFACE)
+target_include_directories(fair-picoscope INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}> $<INSTALL_INTERFACE:include/>)
 
-target_link_libraries(fair-picoscope4000a ps4000a fair-picoscope refl-cpp fmt)
-set_target_properties(gr-digitizers PROPERTIES PUBLIC_HEADER picoscope4000a.hpp)
+target_link_libraries(fair-picoscope INTERFACE ps4000a gr-digitizers-options graph-prototype refl-cpp fmt)
+set_target_properties(gr-digitizers PROPERTIES PUBLIC_HEADER "picoscope.hpp;picoscope4000a.hpp")
 
 if (ENABLE_GR_DIGITIZERS_TESTING)
     add_subdirectory(test)
